Joining Code Cleanup Forces here! CCF FTW

+++ src/main/java/net/canarymod/api/world/position/Position.java
@@ -4,7 +4,7 @@ import net.canarymod.ToolBox;
 
 /**
  * Position
- * 
+ *
  * @author Chris (damagefilter)
  */
 public class Position {
@@ -17,15 +17,15 @@ public class Position {
     }
 
     public Position(int x, int y, int z) {
-        this.x = (double) x;
-        this.y = (double) y;
-        this.z = (double) z;
+        this.x = x;
+        this.y = y;
+        this.z = z;
     }
 
     public Position(float x, float y, float z) {
-        this.x = (double) x;
-        this.y = (double) y;
-        this.z = (double) z;
+        this.x = x;
+        this.y = y;
+        this.z = z;
     }
 
     public Position() {
@@ -36,7 +36,7 @@ public class Position {
 
     /**
      * Copy constructor copies the primitives
-     * 
+     *
      * @param templ
      */
     public Position(Position templ) {
@@ -47,7 +47,7 @@ public class Position {
 
     /**
      * Retrieve X component of Vector
-     * 
+     *
      * @return double x
      */
     public double getX() {
@@ -56,7 +56,7 @@ public class Position {
 
     /**
      * Returns a reliable block ordinate
-     * 
+     *
      * @return
      */
     public int getBlockX() {
@@ -65,7 +65,7 @@ public class Position {
 
     /**
      * Set x component with native double
-     * 
+     *
      * @param x
      */
     public void setX(double x) {
@@ -74,16 +74,16 @@ public class Position {
 
     /**
      * Set x component with a int2double conversion
-     * 
+     *
      * @param x
      */
     public void setX(int x) {
-        this.x = (double) x;
+        this.x = x;
     }
 
     /**
      * Retrieve Y component of Vector
-     * 
+     *
      * @return double y
      */
     public double getY() {
@@ -92,7 +92,7 @@ public class Position {
 
     /**
      * Returns a reliable block ordinate
-     * 
+     *
      * @return
      */
     public int getBlockY() {
@@ -101,7 +101,7 @@ public class Position {
 
     /**
      * Set y component with native double
-     * 
+     *
      * @param y
      */
     public void setY(double y) {
@@ -110,16 +110,16 @@ public class Position {
 
     /**
      * Set y component with a int2double conversion
-     * 
+     *
      * @param y
      */
     public void setY(int y) {
-        this.y = (double) y;
+        this.y = y;
     }
 
     /**
      * Retrieve Z component of Vector
-     * 
+     *
      * @return double z
      */
     public double getZ() {
@@ -128,7 +128,7 @@ public class Position {
 
     /**
      * Returns a reliable block ordinate
-     * 
+     *
      * @return
      */
     public int getBlockZ() {
@@ -137,7 +137,7 @@ public class Position {
 
     /**
      * Set y component with native double
-     * 
+     *
      * @param z
      */
     public void setZ(double z) {
@@ -146,16 +146,16 @@ public class Position {
 
     /**
      * Set z component with a int2double conversion
-     * 
+     *
      * @param z
      */
     public void setZ(int z) {
-        this.z = (double) z;
+        this.z = z;
     }
 
     /**
      * Checks if another object equals this one
-     * 
+     *
      * @param obj
      * @return whether the other object has the same values for x,y,z
      */
